Removed the category of user query and changed the LLM model for chatbot
