+++ frontend/src/components/Chatbot.tsx
@@ -20,7 +20,6 @@ const Chatbot: React.FC = () => {
   const [inputText, setInputText] = useState('');
   const [isLoading, setIsLoading] = useState(false);
   
-  // Add the missing serviceReady state declaration
   const [serviceReady, setServiceReady] = useState(false);
   
   const [chatbotService] = useState(() => new EnhancedChatbotService());
@@ -62,7 +61,6 @@ const Chatbot: React.FC = () => {
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
-    // Add serviceReady check to prevent processing when service isn't ready
     if (!inputText.trim() || isLoading || !serviceReady) return;
 
     const userMessage = inputText.trim();
@@ -71,14 +69,12 @@ const Chatbot: React.FC = () => {
     setIsLoading(true);
 
     try {
-      // Use the enhanced chatbot service with correct method name
       const result = await chatbotService.classifyAndRespond(
         userMessage,
-        'default_user', // In production, use actual user ID
-        'default_session' // In production, use actual session ID
+        'default_user',
+        'default_session'
       );
       
-      // Add bot response with classification metadata
       addMessage(result.response, false, result.intent, result.confidence);
 
     } catch (error) {
@@ -126,11 +122,7 @@ const Chatbot: React.FC = () => {
                   <span className="message-timestamp">
                     {message.timestamp.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}
                   </span>
-                  {message.intent && message.confidence && (
-                    <span className={`intent-debug confidence-${getConfidenceLevel(message.confidence)}`}>
-                      {message.intent} ({(message.confidence * 100).toFixed(0)}%)
-                    </span>
-                  )}
+                  {/* Debug information completely removed from user interface */}
                 </div>
               </div>
             ))}
@@ -179,7 +171,6 @@ const Chatbot: React.FC = () => {
 
 // Helper function to format messages
 const formatMessage = (text: string) => {
-  // Handle numbered lists and formatting
   const lines = text.split('\n');
   return lines.map((line, index) => {
     if (line.match(/^\d+\)/)) {
